+++ app/src/main/java/nextstep/shoppingcart/cart/model/CartViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import nextstep.shoppingcart.Cart
+import nextstep.shoppingcart.data.Cart
 import nextstep.shoppingcart.model.Product
