+++ app/src/main/java/nextstep/shoppingcart/productlist/component/ProductListContents.kt
@@ -6,7 +6,7 @@ import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import nextstep.shoppingcart.ProductsTestData
+import nextstep.shoppingcart.data.ProductsTestData
 import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.component.Product
